Refactor requestDataProcessor for File out into its own file
