+++ lib/resources/Files.js
@@ -1,68 +1,15 @@
 'use strict';
 
-const utils = require('../utils');
-const multipartDataGenerator = require('../MultipartDataGenerator');
-const {StripeError} = require('../Error');
+const {multipartRequestDataProcessor} = require('../multipart');
 const StripeResource = require('../StripeResource');
 const stripeMethod = StripeResource.method;
-
-class StreamProcessingError extends StripeError {}
 
 module.exports = StripeResource.extend({
   path: 'files',
 
   includeBasic: ['list', 'retrieve'],
 
-  requestDataProcessor(method, data, headers, callback) {
-    data = data || {};
-
-    if (method === 'POST') {
-      return getProcessorForSourceType(data);
-    } else {
-      return callback(null, utils.stringifyRequestData(data));
-    }
-
-    function getProcessorForSourceType(data) {
-      const isStream = utils.checkForStream(data);
-      if (isStream) {
-        return streamProcessor(multipartDataGenerator);
-      } else {
-        const buffer = multipartDataGenerator(method, data, headers);
-        return callback(null, buffer);
-      }
-    }
-
-    function streamProcessor(fn) {
-      const bufferArray = [];
-      data.file.data
-        .on('data', (line) => {
-          bufferArray.push(line);
-        })
-        .once('end', () => {
-          const bufferData = Object.assign({}, data);
-          bufferData.file.data = Buffer.concat(bufferArray);
-          const buffer = fn(method, bufferData, headers);
-          callback(null, buffer);
-        })
-        .on('error', (err) => {
-          const errorHandler = streamError(callback);
-          errorHandler(err);
-        });
-    }
-
-    function streamError(callback) {
-      return (error) => {
-        callback(
-          new StreamProcessingError({
-            message:
-              'An error occurred while attempting to process the file for upload.',
-            detail: error,
-          }),
-          null
-        );
-      };
-    }
-  },
+  requestDataProcessor: multipartRequestDataProcessor,
 
   create: stripeMethod({
     method: 'POST',
